Rename package and bump version
