+++ example/example2.py
@@ -1,4 +1,4 @@
-import hawkcatcher
+import hawk_catcher
 import os
 from dotenv import load_dotenv
 
@@ -18,13 +18,13 @@ class Module:
         self.divide_by_zero()
 
     def mannual_sending(self):
-        hawkcatcher.send(ValueError("lol"), {"ping": "pong", "number": 1})
+        hawk_catcher.send(ValueError("lol"), {"ping": "pong", "number": 1})
 
     def send_custom_error(self):
         raise InvalidToken()
 
     def send_with_user(self):
-        hawkcatcher.send(
+        hawk_catcher.send(
             ValueError("USER"),
             None,
             {'id': 1, 'name': 'Alice'}
@@ -35,9 +35,9 @@ def main():
     token = os.getenv('HAWK_TOKEN')
 
     if token is None or token == "":
-        print('hawkcatcher token not provided. Please provide HAWK_TOKEN variable in .env file')
+        print('hawk-catcher token not provided. Please provide HAWK_TOKEN variable in .env file')
         return
-    hawkcatcher.init(token)
+    hawk_catcher.init(token)
     test = Module()
     test.mannual_sending()
     test.send_with_user()
