+++ src/hawk_catcher/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "3.5.0"
+__version__ = "3.5.1"
 
 from .core import Hawk
 from .types import HawkCatcherSettings
