+++ src/hawk_catcher/errors.py
