+++ src/hawk_catcher/modules/fastapi/fastapi.py
@@ -1,11 +1,11 @@
-from hawkcatcher.types import HawkCatcherSettings
+from hawk_catcher.types import HawkCatcherSettings
 from ...core import Hawk
-from hawkcatcher.modules.fastapi.types import FastapiSettings, FastapiAddons
+from hawk_catcher.modules.fastapi.types import FastapiSettings, FastapiAddons
 from starlette.types import ASGIApp, Receive, Scope, Send
 from starlette.requests import Request
 from starlette.middleware.base import BaseHTTPMiddleware
 from typing import Union
-from hawkcatcher.errors import ModuleError
+from hawk_catcher.errors import ModuleError
 import asyncio
 from contextvars import ContextVar
 from fastapi import Request
